bugs #12679 fix(archive-unit): other metadata in display and edit mode bugs #12701 fix(archive-unit): missing metadata in display and edit mode
changes:
* (style) apply correct shadow to accordion header
* (clean-code) fix usage of 'end' to 'flex-end'
* (clean-code) remove console logs
* (refactor) rationalization and optimisation for display and edit mode
