+++ ui/ui-frontend-common/src/app/modules/archive-unit/components/archive-unit-editor/archive-unit-editor.component.html
@@ -1,0 +1,1 @@
+<vitamui-common-object-editor [editObject]="editObject$ | async"></vitamui-common-object-editor>
