+++ ui/ui-frontend-common/src/app/modules/archive-unit/components/archive-unit-editor/components/editor-banner/editor-banner.component.scss
@@ -20,5 +20,5 @@
 .actions {
   display: flex;
   align-items: center;
-  justify-content: end;
+  justify-content: flex-end;
 }
