+++ ui/ui-frontend-common/src/app/modules/archive-unit/components/archive-unit-editor/components/editor-banner/editor-banner.component.ts
