+++ ui/ui-frontend-common/src/app/modules/archive-unit/components/archive-unit-viewer/archive-unit-viewer.component.spec.ts
@@ -1,6 +1,7 @@
 import { HttpClientTestingModule } from '@angular/common/http/testing';
 import { CUSTOM_ELEMENTS_SCHEMA, SimpleChange } from '@angular/core';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { ReactiveFormsModule } from '@angular/forms';
 import { TranslateModule } from '@ngx-translate/core';
 import { Observable } from 'rxjs';
 import { map } from 'rxjs/operators';
@@ -9,8 +10,8 @@ import { BASE_URL } from '../../../injection-tokens';
 import { LoggerModule } from '../../../logger';
 import { Collection, Schema } from '../../../models';
 import { ObjectViewerModule } from '../../../object-viewer/object-viewer.module';
-import { MockSchemaService } from '../../../object-viewer/services/mock-schema.service';
 import { SchemaService } from '../../../schema';
+import { MockSchemaService } from '../../../schema/mock-schema.service';
 import { ArchiveUnitViewerComponent } from './archive-unit-viewer.component';
 
 describe('ArchiveUnitViewerComponent', () => {
@@ -64,7 +65,7 @@ describe('ArchiveUnitViewerComponent', () => {
     await TestBed.configureTestingModule({
       declarations: [ArchiveUnitViewerComponent],
       schemas: [CUSTOM_ELEMENTS_SCHEMA],
-      imports: [HttpClientTestingModule, ObjectViewerModule, LoggerModule.forRoot(), TranslateModule.forRoot()],
+      imports: [HttpClientTestingModule, ObjectViewerModule, ReactiveFormsModule, LoggerModule.forRoot(), TranslateModule.forRoot()],
       providers: [
         SchemaService,
         { provide: BASE_URL, useValue: '/fake-api' },
