+++ ui/ui-frontend-common/src/app/modules/archive-unit/components/physical-archive-viewer/physical-archive-viewer.component.ts
