+++ ui/ui-frontend-common/src/app/modules/components/accordion/accordion.component.html
@@ -1,4 +1,4 @@
-<div class="header text bold mat-elevation-z1" [ngClass]="{ opened: opened }" (click)="toggle()">
+<div class="header text bold elevation-1-dark" [ngClass]="{ opened: opened }" (click)="toggle()">
   <i class="vitamui-icon icon" [ngClass]="{ 'vitamui-icon-chevron-down': opened, 'vitamui-icon-chevron-right': !opened }"></i>
   <div class="v-row">
     <div class="v-col">
