+++ ui/ui-frontend-common/src/app/modules/components/accordion/accordion.component.scss
@@ -1,3 +1,5 @@
+@import '../../../../sass/mixins/elevation';
+
 :host {
   width: 100%;
 }
@@ -26,6 +28,9 @@
 }
 
 .header {
+  box-shadow:
+    0 0 2px 0 rgb(33 33 33 / 20%),
+    0 1px 4px 0 rgb(33 33 33 / 5%);
   min-height: 50px;
   padding: 0px 16px;
   border-radius: 5px;
