+++ ui/ui-frontend-common/src/app/modules/index.ts
@@ -52,15 +52,15 @@ export * from './app-root-component.class';
 export * from './app.guard';
 export * from './application-id.enum';
 export * from './application.service';
-export * from './archive/index';
+export * from './archive-unit/index';
 export * from './auth.guard';
 export * from './auth.service';
 export { AuthenticationModule } from './authentication/authentication.module';
-export * from './vitamui-global-error-handler';
-export * from './vitamui-table/index';
+export * from './country.service';
 export * from './error-dialog/error-dialog.component';
 export * from './global-event.service';
 export * from './injection-tokens';
+export * from './language.service';
 export * from './logbook/index';
 export * from './module-import-guard';
 export * from './paginated-api.interface';
@@ -72,20 +72,20 @@ export * from './startup.service';
 export * from './subrogation/index';
 export * from './tenant-selection.guard';
 export * from './theme.service';
-export * from './language.service';
-export * from './country.service';
 export * from './utils/index';
+export * from './vitamui-global-error-handler';
+export * from './vitamui-table/index';
 
-export * from './components/accordion/accordion.component';
-export * from './components/data/data.component';
-export * from './components/application/index';
 export * from './base-http-client';
+export * from './components/accordion/accordion.component';
 export * from './components/application-select-content/index';
 export * from './components/application/index';
+export * from './components/autocomplete';
 export * from './components/blank/blank.component';
 export * from './components/cancelled-snack-bar/index';
 export * from './components/collapse/collapse.module';
 export * from './components/confirm-dialog/index';
+export * from './components/data/data.component';
 export * from './components/download-snack-bar/index';
 export * from './components/editable-field/index';
 export * from './components/header/select-tenant-dialog/index';
@@ -96,7 +96,6 @@ export * from './components/search-bar/index';
 export * from './components/stepper/index';
 export * from './components/table-filter/index';
 export * from './components/user-alerts/index';
-export * from './components/autocomplete';
 export * from './components/vitamui-customer-select/index';
 export * from './components/vitamui-display-node/index';
 export * from './components/vitamui-field-error/index';
