+++ ui/ui-frontend-common/src/app/modules/object-viewer/components/primitive/primitive.component.html
@@ -1,1 +1,1 @@
-<vitamui-common-data [label]="displayObject.displayRule.ui.label" [value]="value"></vitamui-common-data>
+<vitamui-common-data [label]="displayObject.displayRule.ui.label | translate" [value]="value"></vitamui-common-data>
