+++ ui/ui-frontend-common/src/app/modules/object-viewer/models/schema-element.model.ts
@@ -44,7 +44,7 @@ export interface SchemaElement extends WithPath {
   readonly ApiField: string;
   readonly Type: 'TEXT' | 'KEYWORD' | 'DATE' | 'OBJECT' | 'BOOLEAN' | 'LONG';
   readonly DataType?: 'STRING' | 'DATETIME' | 'OBJECT' | 'LONG' | 'DOUBLE' | 'BOOLEAN';
-  readonly Origin: string;
+  readonly Origin: 'INTERNAL' | 'EXTERNAL' | 'VIRTUAL';
   readonly SedaField?: string;
   readonly ShortName?: string;
   readonly Description?: string;
