+++ ui/ui-frontend/projects/archive-search/src/app/archive/archive-preview/archive-unit-description-tab/archive-unit-description-tab.component.ts
@@ -97,7 +97,10 @@ export class ArchiveUnitDescriptionTabComponent implements OnDestroy {
         .pipe(
           filter((result) => !!result),
           map(() => this.archiveUnitEditor.getJsonPatch()),
-          tap(() => this.spinnerOverlayService.open()),
+          tap(
+            () => this.spinnerOverlayService.open(),
+            (err) => this.logger.error(this, err),
+          ),
           switchMap((jsonPatchDto) => this.archiveUnitService.asyncPartialUpdateArchiveUnitByCommands(jsonPatchDto)),
         )
         .subscribe(
